# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateWaitlistEntry {
  neighborhood: string;
  phone: string;
}

export interface UpdateWaitlistName {
  name: string;
}

export interface WaitlistEntry {
  id: number;
  neighborhood: string;
  phone: string;
  name?: string | null;
  createdAt: string;
}

export interface TrackPageView {
  page: string;
}

export interface PageView {
  id: number;
  page: string;
  createdAt: string;
}

export interface AnalyticsSummary {
  totalPageViews: number;
  totalSignups: number;
  signupsWithName: number;
  conversionRate: number;
}

export interface ErrorResponse {
  message: string;
}